requirements.txt
fastapi~=0.72.0
opentelemetry-exporter-jaeger~=1.8.0
opentelemetry-instrumentation-fastapi~=0.27b0
starlette~=0.17.1
uvicorn~=0.17.0
